Rename File whit lower identifier
