+++ analysis_options.yaml
@@ -1,1 +1,4 @@
+analyzer:
+  errors:
+    no_leading_underscores_for_local_identifiers: ignore
 include: package:flutter_lints/flutter.yaml
